ref: move "methods" for each provider type into their own dirs
