--- src/server/models/listItem/listItemService/types.ts
@@ -1,11 +1,0 @@
-export interface ListItemService {
-  getPublished: () => {};
-  checkIfExists: () => {};
-  getAllListItems: () => {};
-  findById: () => {};
-  toggleApprovalStatus: () => {};
-  togglePublishedStatus: () => {};
-  deleteListItem: () => {};
-  findItemsByCountry: () => {};
-  setEmailIsVerified: () => {};
-}
