+++ src/server/models/listItem/providers/index.ts
@@ -1,0 +1,2 @@
+export * as LawyerListItem from "./Lawyers";
+export * as CovidTestSupplierListItem from "./CovidTestSupplier";
