+++ src/server/models/listItem/providers/types.ts
@@ -1,0 +1,7 @@
+import { Address, Country, ListItem } from "server/models/types";
+
+export interface ListItemWithAddressCountry extends ListItem {
+  address: Address & {
+    country: Country;
+  };
+}
